Add Lung Segmentation (Hounsfield based) (closes #120)
* add lung segmentation

* catch InvalidDicomError properly
