+++ prediction/requirements/local.txt
@@ -1,3 +1,4 @@
 -r base.txt
 flake8==3.3.0
 pytest==3.1.3
+pylidc==0.1.8
