+++ prediction/src/algorithms/identify/helpers.py
@@ -1,42 +1,5 @@
-import glob
-import logging
-import os
-from collections import defaultdict
-
 import cv2
-import dicom
 import numpy
-import scipy
-from skimage.filters import roberts
-from skimage.measure import label, regionprops
-from skimage.morphology import disk, binary_erosion, binary_closing
-from skimage.segmentation import clear_border
-
-PATIENT_LIST = None
-PRINT_TAB_MAP = defaultdict(lambda: [])
-
-
-def load_patient_images(patient_id, base_dir=None, wildcard="*.*", exclude_wildcards=None):
-    exclude_wildcards = exclude_wildcards or []
-    src_dir = base_dir + patient_id + "/"
-    src_img_paths = glob.glob(src_dir + wildcard)
-    for exclude_wildcard in exclude_wildcards:
-        exclude_img_paths = glob.glob(src_dir + exclude_wildcard)
-        src_img_paths = [im for im in src_img_paths if im not in exclude_img_paths]
-    src_img_paths.sort()
-    images = [cv2.imread(img_path, cv2.IMREAD_GRAYSCALE) for img_path in src_img_paths]
-    images = [im.reshape((1,) + im.shape) for im in images]
-    res = numpy.vstack(images)
-    return res
-
-
-def normalize_hu(image):
-    MIN_BOUND = -1000.0
-    MAX_BOUND = 400.0
-    image = (image - MIN_BOUND) / (MAX_BOUND - MIN_BOUND)
-    image[image > 1] = 1.
-    image[image < 0] = 0.
-    return image
 
 
 def rescale_patient_images(images_zyx, org_spacing_xyz, target_voxel_mm, is_mask_image=False):
@@ -101,75 +64,3 @@ def rescale_patient_images2(images_zyx, target_shape):
     res = res.swapaxes(0, 2)
     res = res.swapaxes(2, 1)
     return res
-
-
-def get_segmented_lungs(im):
-    # Step 1: Convert into a binary image.
-    binary = im < -400
-    # Step 2: Remove the blobs connected to the border of the image.
-    cleared = clear_border(binary)
-    # Step 3: Label the image.
-    label_image = label(cleared)
-    # Step 4: Keep the labels with 2 largest areas.
-    areas = [r.area for r in regionprops(label_image)]
-    areas.sort()
-    if len(areas) > 2:
-        for region in regionprops(label_image):
-            if region.area < areas[-2]:
-                for coordinates in region.coords:
-                    label_image[coordinates[0], coordinates[1]] = 0
-    binary = label_image > 0
-    # Step 5: Erosion operation with a disk of radius 2.
-    # This operation is seperate the lung nodules attached to the blood vessels.
-    selem = disk(2)
-    binary = binary_erosion(binary, selem)
-    # Step 6: Closure operation with a disk of radius 10.
-    # This operation is to keep nodules attached to the lung wall.
-    selem = disk(10)  # CHANGE BACK TO 10
-    binary = binary_closing(binary, selem)
-    # Step 7: Fill in the small holes inside the binary mask of lungs.
-    edges = roberts(binary)
-    binary = scipy.ndimage.binary_fill_holes(edges)
-    # Step 8: Superimpose the binary mask on the input image.
-    get_high_vals = binary == 0
-    im[get_high_vals] = -2000
-    return im, binary
-
-
-def load_patient(src_dir):
-    slices = []
-    for s in os.listdir(src_dir):
-        try:
-            slices.append(dicom.read_file(src_dir + '/' + s))
-        except Exception as e:
-            logging.error("{} no valid DICOM: {}".format(s, e))
-    slices.sort(key=lambda x: int(x.InstanceNumber))
-    try:
-        slice_thickness = numpy.abs(slices[0].ImagePositionPatient[2] - slices[1].ImagePositionPatient[2])
-    except IndexError as e:
-        slice_thickness = numpy.abs(slices[0].SliceLocation - slices[1].SliceLocation)
-
-    for s in slices:
-        s.SliceThickness = slice_thickness
-    return slices
-
-
-def get_pixels_hu(slices):
-    image = numpy.stack([s.pixel_array for s in slices])
-    image = image.astype(numpy.int16)
-    image[image == -2000] = 0
-    for slice_number in range(len(slices)):
-        intercept = slices[slice_number].RescaleIntercept
-        slope = slices[slice_number].RescaleSlope
-        if slope != 1:
-            image[slice_number] = slope * image[slice_number].astype(numpy.float64)
-            image[slice_number] = image[slice_number].astype(numpy.int16)
-        image[slice_number] += numpy.int16(intercept)
-
-    return numpy.array(image, dtype=numpy.int16)
-
-
-def cv_flip(img, cols, rows, degree):
-    M = cv2.getRotationMatrix2D((cols / 2, rows / 2), degree, 1.0)
-    dst = cv2.warpAffine(img, M, (cols, rows))
-    return dst
